+++ wsi_service_base_plugins/pil/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.3.0
+
+- Added support for regions outside of the image
+
 ## 0.2.0
 
 - Updated to v3 models